// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/DetailedAgreementDTO.ts
export const DetailedAgreementDTOStatusEnum = {
    Active: 'ACTIVE',
    Closed: 'CLOSED',
    Unknown: 'UNKNOWN'
} as const;
export type DetailedAgreementDTOStatusEnum = typeof DetailedAgreementDTOStatusEnum[keyof typeof DetailedAgreementDTOStatusEnum];