// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/BalanceWithAgreementStatusDTO.ts
export interface BalanceWithAgreementStatusDTO {
    /**
     * 
     * @type {string}
     * @memberof BalanceWithAgreementStatusDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof BalanceWithAgreementStatusDTO
     */
    agreementId?: string;
    /**
     * 
     * @type {string}
     * @memberof BalanceWithAgreementStatusDTO
     */
    currency?: string;
    /**
     * 
     * @type {Date}
     * @memberof BalanceWithAgreementStatusDTO
     */
    calculationDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BalanceWithAgreementStatusDTO
     */
    latestEventDate?: Date;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof BalanceWithAgreementStatusDTO
     */
    values?: { [key: string]: number; };
    /**
     * 
     * @type {string}
     * @memberof BalanceWithAgreementStatusDTO
     */
    balanceEventId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BalanceWithAgreementStatusDTO
     */
    agreementActive?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof BalanceWithAgreementStatusDTO
     */
    activationDate?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof BalanceWithAgreementStatusDTO
     */
    repaid?: boolean;
}