// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/RepaymentScheduleResponse.ts
export interface RepaymentScheduleResponse {
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    agreementId?: string;
    /**
     * 
     * @type {Date}
     * @memberof RepaymentScheduleResponse
     */
    creationDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof RepaymentScheduleResponse
     */
    scheduledDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    scheduleType?: RepaymentScheduleResponseScheduleTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    repaymentPhaseType?: RepaymentScheduleResponseRepaymentPhaseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    repaymentStatus?: RepaymentScheduleResponseRepaymentStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    amount?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    principal?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    interest?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    fees?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    remainingAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    paymentDaysOverdue?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    chargedLateFeesAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    errorCode?: string;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    errorMessage?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof RepaymentScheduleResponse
     */
    externalIds?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    overpayment?: number;
    /**
     * 
     * @type {number}
     * @memberof RepaymentScheduleResponse
     */
    overdueAmount?: number;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    phaseType?: RepaymentScheduleResponsePhaseTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof RepaymentScheduleResponse
     */
    type?: RepaymentScheduleResponseTypeEnum;
}


/**
 * @export
 */
export const RepaymentScheduleResponseScheduleTypeEnum = {
    Base: 'BASE',
    Additional: 'ADDITIONAL',
    EarlyRepayment: 'EARLY_REPAYMENT',
    InterestRateBased: 'INTEREST_RATE_BASED',
    Rebate: 'REBATE',
    RevenueBased: 'REVENUE_BASED',
    DailyPayoutBased: 'DAILY_PAYOUT_BASED',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentScheduleResponseScheduleTypeEnum = typeof RepaymentScheduleResponseScheduleTypeEnum[keyof typeof RepaymentScheduleResponseScheduleTypeEnum];

/**
 * @export
 */
export const RepaymentScheduleResponseRepaymentPhaseTypeEnum = {
    RepaymentPhase: 'REPAYMENT_PHASE',
    DrawPhase: 'DRAW_PHASE',
    FixedAmount: 'FIXED_AMOUNT',
    ForbearanceFixedAmount: 'FORBEARANCE_FIXED_AMOUNT',
    PostForbearanceFixedAmount: 'POST_FORBEARANCE_FIXED_AMOUNT',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentScheduleResponseRepaymentPhaseTypeEnum = typeof RepaymentScheduleResponseRepaymentPhaseTypeEnum[keyof typeof RepaymentScheduleResponseRepaymentPhaseTypeEnum];

/**
 * @export
 */
export const RepaymentScheduleResponseRepaymentStatusEnum = {
    Scheduled: 'SCHEDULED',
    Pending: 'PENDING',
    PartiallyPaid: 'PARTIALLY_PAID',
    Paid: 'PAID',
    Cancelled: 'CANCELLED',
    Suspended: 'SUSPENDED',
    Merged: 'MERGED',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentScheduleResponseRepaymentStatusEnum = typeof RepaymentScheduleResponseRepaymentStatusEnum[keyof typeof RepaymentScheduleResponseRepaymentStatusEnum];

/**
 * @export
 */
export const RepaymentScheduleResponsePhaseTypeEnum = {
    RepaymentPhase: 'REPAYMENT_PHASE',
    DrawPhase: 'DRAW_PHASE',
    FixedAmount: 'FIXED_AMOUNT',
    ForbearanceFixedAmount: 'FORBEARANCE_FIXED_AMOUNT',
    PostForbearanceFixedAmount: 'POST_FORBEARANCE_FIXED_AMOUNT',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentScheduleResponsePhaseTypeEnum = typeof RepaymentScheduleResponsePhaseTypeEnum[keyof typeof RepaymentScheduleResponsePhaseTypeEnum];

/**
 * @export
 */
export const RepaymentScheduleResponseTypeEnum = {
    Base: 'BASE',
    Additional: 'ADDITIONAL',
    EarlyRepayment: 'EARLY_REPAYMENT',
    InterestRateBased: 'INTEREST_RATE_BASED',
    Rebate: 'REBATE',
    RevenueBased: 'REVENUE_BASED',
    DailyPayoutBased: 'DAILY_PAYOUT_BASED',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentScheduleResponseTypeEnum = typeof RepaymentScheduleResponseTypeEnum[keyof typeof RepaymentScheduleResponseTypeEnum];