// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/CustomerTransactionFeeBalance.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-agreements
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CustomerTransactionFeeBalance
 */
export interface CustomerTransactionFeeBalance {
    /**
     * 
     * @type {string}
     * @memberof CustomerTransactionFeeBalance
     */
    agreementId?: string;
    /**
     * 
     * @type {number}
     * @memberof CustomerTransactionFeeBalance
     */
    billPaymentFees?: number;
    /**
     * 
     * @type {number}
     * @memberof CustomerTransactionFeeBalance
     */
    cardPaymentFees?: number;
    /**
     * 
     * @type {number}
     * @memberof CustomerTransactionFeeBalance
     */
    cashFees?: number;
    /**
     * 
     * @type {number}
     * @memberof CustomerTransactionFeeBalance
     */
    lateFees?: number;
}

/**
 * Check if a given object implements the CustomerTransactionFeeBalance interface.
 */
export function instanceOfCustomerTransactionFeeBalance(value: object): value is CustomerTransactionFeeBalance {
    return true;
}

export function CustomerTransactionFeeBalanceFromJSON(json: any): CustomerTransactionFeeBalance {
    return CustomerTransactionFeeBalanceFromJSONTyped(json, false);
}

export function CustomerTransactionFeeBalanceFromJSONTyped(json: any, ignoreDiscriminator: boolean): CustomerTransactionFeeBalance {
    if (json == null) {
        return json;
    }
    return {
        
        'agreementId': json['agreementId'] == null ? undefined : json['agreementId'],
        'billPaymentFees': json['billPaymentFees'] == null ? undefined : json['billPaymentFees'],
        'cardPaymentFees': json['cardPaymentFees'] == null ? undefined : json['cardPaymentFees'],
        'cashFees': json['cashFees'] == null ? undefined : json['cashFees'],
        'lateFees': json['lateFees'] == null ? undefined : json['lateFees'],
    };
}

export function CustomerTransactionFeeBalanceToJSON(json: any): CustomerTransactionFeeBalance {
    return CustomerTransactionFeeBalanceToJSONTyped(json, false);
}

export function CustomerTransactionFeeBalanceToJSONTyped(value?: CustomerTransactionFeeBalance | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'agreementId': value['agreementId'],
        'billPaymentFees': value['billPaymentFees'],
        'cardPaymentFees': value['cardPaymentFees'],
        'cashFees': value['cashFees'],
        'lateFees': value['lateFees'],
    };
}